Created useWeather hook
