+++ app/(tabs)/profile.tsx
@@ -2,40 +2,12 @@ import React, { useState } from 'react';
 import { View, StyleSheet } from 'react-native';
 import { HeaderModal } from '@/components/Profile/HeaderModal';
 import { WeatherInfo } from '@/components/Profile/HeaderModal/types'; 
+import { useWeather } from '@/hooks/useWeather';
 
-const getCurrentDate = (): string => {
-  const date = new Date();
-  return date.toLocaleDateString('fr-FR', {
-    day: 'numeric',
-    month: 'short'
-  }).replace('.', '');
-};
-
-const generateWeatherDates = () => {
-  const today = new Date();
-  const forecast = [];
-  
-  for (let i = -2; i <= 2; i++) {
-    const date = new Date(today);
-    date.setDate(today.getDate() + i);
-    
-    forecast.push({
-      dayName: date.toLocaleDateString('fr-FR', { weekday: 'short' }).replace('.', '')+ ' ',
-      date: date.toLocaleDateString('fr-FR', { day: 'numeric', month: 'short' }).replace('.', ''),
-      temp: Math.floor(Math.random() * (25 - 15) + 15),
-      minTemp: Math.floor(Math.random() * (15 - 8) + 8)
-    });
-  }
-
-  return {
-    currentDate: getCurrentDate(),
-    forecast: forecast
-  };
-};
 
 export default function ProfileScreen () {
   const [variant, setVariant] = useState<'private' | 'public'>('private');
-  const weatherData = generateWeatherDates();
+  const { weatherData } = useWeather();
 
 
   const handleToggleVariant = () => {
@@ -48,7 +20,7 @@ export default function ProfileScreen () {
         variant={variant}
         username="Dina Hsisou"
         location="Ouarzazat"
-        weather={weatherData}
+        weather={weatherData ?? undefined}
         onToggleVariant={handleToggleVariant}
         onCalendarPress={() => console.log('Calendar pressed')}
         onSettingsPress={() => console.log('Settings pressed')}
