+++ components/Profile/HeaderModal/styles.ts
@@ -238,4 +238,5 @@ export const styles = StyleSheet.create({
     fontSize: 16,
     color: '#000',
   },
+
 });
